requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
decorator==4.4.2
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instagrapi==2.2.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
moviepy==1.0.3
narwhals==2.15.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
proglog==0.1.12
protobuf==6.33.5
pyarrow==23.0.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.4
rpds-py==0.30.0
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
streamlit==1.53.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
